# aspnetcore/src/IO.Swagger/Controllers/DigitalGoodsApi.cs
/*
 * Shipments API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class DigitalGoodsApiController : Controller
    { 
        /// <summary>
        /// Digital Goods Fulfillment Status Update. (Auth policies: OrderFulfiller)
        /// </summary>
        
        /// <param name="partnerOrderNumber">The partner order number.</param>
        /// <param name="costcoTrackingId">Unique ID to Correlate HTTP requests between a client and server.</param>
        /// <param name="lineItems">The collection of items to process</param>
        /// <param name="costcoEnv">Non-Production Environment Identifier.</param>
        /// <response code="200">Success</response>
        /// <response code="400">Bad Request</response>
        /// <response code="500">Server Error</response>
        [HttpPut]
        [Route("/shipments/v1/fulfillment/digitalgoods/orders/{partnerOrderNumber}")]
        [ValidateModelState]
        [SwaggerOperation("FulfillDigitalGoods")]
        [SwaggerResponse(statusCode: 400, type: typeof(ApiExceptionResponse), description: "Bad Request")]
        [SwaggerResponse(statusCode: 500, type: typeof(ApiExceptionResponse), description: "Server Error")]
        public virtual IActionResult FulfillDigitalGoods([FromRoute][Required]string partnerOrderNumber, [FromHeader][Required()]string costcoTrackingId, [FromBody]List<LineItem> lineItems, [FromHeader]string costcoEnv)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ApiExceptionResponse));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(ApiExceptionResponse));


            throw new NotImplementedException();
        }
    }
}