# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "Shipments API"
  },
  "host" : "api.costco.com",
  "basePath" : "/shipments/v1",
  "schemes" : [ "https" ],
  "security" : [ {
    "Bearer" : [ ]
  } ],
  "paths" : {
    "/fulfillment/digitalgoods/orders/{partnerOrderNumber}" : {
      "put" : {
        "tags" : [ "DigitalGoods" ],
        "summary" : "Digital Goods Fulfillment Status Update. (Auth policies: OrderFulfiller)",
        "operationId" : "FulfillDigitalGoods",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "name" : "partnerOrderNumber",
          "in" : "path",
          "description" : "The partner order number.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "lineItems",
          "description" : "The collection of items to process",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/LineItem"
            },
            "uniqueItems" : false
          }
        }, {
          "name" : "costco.trackingId",
          "in" : "header",
          "description" : "Unique ID to Correlate HTTP requests between a client and server.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "costco.env",
          "in" : "header",
          "description" : "Non-Production Environment Identifier.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/ApiExceptionResponse"
            }
          },
          "500" : {
            "description" : "Server Error",
            "schema" : {
              "$ref" : "#/definitions/ApiExceptionResponse"
            }
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "Bearer" : {
      "description" : "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "type" : "apiKey",
      "name" : "authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "LineItem" : {
      "type" : "object",
      "required" : [ "fulfilledOnTimeStamp", "fulfillmentDescription", "fulfillmentStatus", "itemNumber", "quantity" ],
      "properties" : {
        "itemNumber" : {
          "type" : "string",
          "description" : "The Costco item number",
          "minLength" : 1
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "The quantity of units to be fulfilled",
          "minimum" : 1,
          "maximum" : 1000
        },
        "fulfilledOnTimeStamp" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "The moment at which the order fulfillment status changed in ISO 8601 format."
        },
        "fulfillmentStatus" : {
          "type" : "string",
          "description" : "An explanation for the fulfillment status change",
          "enum" : [ "delivered", "cancelled" ]
        },
        "fulfillmentDescription" : {
          "type" : "string",
          "description" : "An explanation for the fulfillment status change",
          "minLength" : 1
        }
      },
      "description" : "LineItem for an order"
    },
    "ApiExceptionResponse" : {
      "type" : "object",
      "required" : [ "trackingId" ],
      "properties" : {
        "errorType" : {
          "type" : "string",
          "readOnly" : true
        },
        "userMessage" : {
          "type" : "string",
          "readOnly" : true
        },
        "developerMessage" : {
          "type" : "string",
          "readOnly" : true
        },
        "errorCode" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        },
        "source" : {
          "type" : "string",
          "readOnly" : true
        },
        "trackingId" : {
          "type" : "string"
        }
      }
    }
  }
}